% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{makeForest}
\alias{makeForest}
\title{Make Forest Plot}
\usage{
makeForest(catmapobject, summary = "", main = "Main Title",
  mean = exp(catmapobject$logOR), lower = catmapobject$lbci.fe,
  upper = catmapobject$ubci.fe, study = c("Study", sub(",", " ",
  catmapobject$studyname)))
}
\arguments{
\item{catmapobject}{A catmap object created by \code{\link{catmap}}.}

\item{summary}{A character string. The kind of summary statistic
to plot. Select from "fixed" or "random".}

\item{main}{A character string. The figure title.}

\item{mean, lower, upper, study}{Numeric or character vectors.
Used to guide the construction of the forest plot.}
}
\description{
A back-end wrapper function used to make forest plots.
}
